Only run Pharos when the pull request is ready for review. (#128)
Fixes an issue in which the Pharos workflow would crash for draft PRs, as draft PRs are configured to not push Docker containers in the build stage. This PR turns the Pharos workflow off for draft PRs, but runs the Pharos workflow automatically when the PR is marked as `ready for review`.